+++ source/arbok/include/arbok/gabow/active_forest.h
@@ -10,10 +10,10 @@ namespace arbok {
 
     class ActiveForest {
     public:
-        explicit ActiveForest(CompressedTree<int>& _co);
+        explicit ActiveForest(CompressedTree<double>& _co);
         ~ActiveForest();
 
-        void makeActive(int from, int to, int weight, int id);
+        void makeActive(int from, int to, double weight, int id);
         void deleteActiveEdge(int i);
         int getMin(int i); // returns min from heap of node i, does NOT delete it
         void mergeHeaps(int i, int j);
@@ -24,9 +24,9 @@ namespace arbok {
         void removeFromCurrentList(FibHeapNode* v);
         void moveHome(FibHeapNode* v);
         void loseChild(FibHeapNode* v);
-        int curWeight(FibHeapNode* v);
+        double curWeight(FibHeapNode* v);
 
-        CompressedTree<int>& co;
+        CompressedTree<double>& co;
         std::vector<FibHeapNode*> active_edge; // for each node the active outgoing edge
         std::vector<std::list<FibHeapNode*>> active_sets; // for each node on path the active set heap represented by the root list
     };
